requirements.txt
streamlit
youtube-transcript-api
openai
python-dotenv
yt-dlp
torch